Eliminate using separate publishconf.py
To get absolute links, it was needed to use SITEURL but with the way
the import was done, pelicanconf.py did not have access to what
publishconf.py set it to.  I made publishconf.py just a symlink to
pelicanconf.py and then determine the right value by examining
__file__.  Yes, it is hacky.

This commit also has extensive changes to pelicanconf.py to organize
and document it better.  Finally, EXTRA_PATH_METADATA is now generated
automatically.