Update most images that did not have known provenance
There are not that many images on the web site, but the provenance of
many of them was unknown.

The various images of antennas and cables on the HDTV Antennas page
have all been replaced with pictures that I have personally taken and
have released with the CC0 license.

The website's logo (the image in the oval at the top right of every
page) and the favicon have been replaced with a scan of watercolor
painted by Kim Miller that she has graciously allowed us to use:
  Please find attached a copy of my painting “Heceta morning by the Sea”.
  I hereby give you the right to use it for the Western Lane TV website
  and you may alter it in any way you have a need to.

The Siuslaw News previously given us permission to use the newspaper
images on the History page with the following credit:
  Courtesy of The Siuslaw News www.thesiuslawnews.com

The image on the HDTV Coverage map was generated by me via a website
and has the appropriate credits embedded in the image.

The only other image on the site is the picture of the tower on the
Leasing page.  I have to assume that whoever took it must be a member
and has allowed its presence on the website.